created visualize separate file
